core/internal: discard errors when analytics cannot be sent
Currently, when analytics cannot be sent to 'chichi.open2b.net',
Meergo logs a lot of errors into the stderr.

We do not expect this behavior, as it may confuse the user of Meergo
with information they don't care about.

So, this commit passes a 'discardLogger' to Analytics so that log and
error messages are discarded.

Resolves #1887.