core/internals/transformers: refine validation error messages
This change refines the error messages produced during validation of
values returned by transformation functions.

Previously, when using Python, error messages referred to invalid fields
as “key”, even though the validation operates on schema properties. The
term `key` was also misleading, as it could be mistaken for an 
interpreter-related Python error.

Now, when a returned property is not defined in the output schema, the
message changes from:

> key «first_name» does not exist

to:

> property «first_name» is not part of output schema; rename or remove
  it in the transformation function