all: load images from external URLs and remove images from Meergo
Motivation of this commit: it is necessary that the icons are no longer
stored within the repository of Meergo, and they must be stored instead
in an external position.

Thus, this commit makes the Admin console and the docs load the images
from an external source, then removes the icons from this repository.

It adds a new environment variable 'MEERGO_EXTERNAL_ASSETS_URLS', which
contains a list of base URLs from which the icons can be retrieved from
the Admin console.

Co-authored-by: Andrea Retini <retini@open2b.com>