test: fix and update `TestReimport`
Occasionally, the test 'TestReimport' fails.

This is probably caused by a temporization problem that involves the
execution of the Identity Resolution and the test execution. Maybe that
test was written before we had the APIs for the Identity Resolution.

So, this commit makes the test control explicitly the Identity
Resolution. This should fix the occasional failing.

Moreover, this commit updates the test by removing references to an old
issue, enabling some disabled checks, and generally fixing some
comments.