core/internal/collector: read user agent only from event context
Previously, when an event was ingested and `context.userAgent` was
missing, the request's user agent was read and parsed to fill
`context.browser`, `context.os`, and `context.userAgent`.

This approach was incorrect because the request's user agent might not
come from a browser, and only a browser user agent can provide reliable 
information about the browser and operating system.

In addition, the request's user agent is irrelevant when the request
does not originate from the user's device, such as when it comes from
server-side SDKs.

This change removes the use of the request's user agent and relies
solely on the user agent provided in the event context.