all: define per-connector default behavior for request IP fallback
Before this change, when an event was received and no `context.ip` was
provided, the request IP was always used as the event IP (except for the
special case `0.0.0.0`).

With this change, each connector can specify its desired default
behavior through the `SDKInfo.FallbackToRequestIP` field. For JavaScript
and Android connectors, the default is to fall back to the request IP;
for all other connectors, it is not.

This change also adds three new special IP values that allow forcing the
use of the request IP, with or without obfuscation.