cmd: clear Meergo env vars, except those for connectors
Currently, the Meergo environment variables are kept after reading the
configuration so the connectors can access them at any moment.

This commit changes this mechanism by unsetting the Meergo environment
variables, except for those intended for connectors. This is feasible
because the connector env vars all have the prefix "MEERGO_CONNECTOR_".

This reduces the possibility that any point in the code can read the
configuration passed from the environment.

Resolves #1912.