meergo: run local transf. functions with different user in Docker
This commit ensures that:

(1) A non-root user is created during the Docker build.
(2) Docker provides this user as the user to run local Meergo
transformation functions.

This reduces the privileges of the process running the JavaScript and
Python functions, thus preventing them from causing significant damage
to the container filesystem.

Obviously, this isn't intended for security, as the local transformer is
not considered secure anyway, but it increases usability and prevents
unexpected behavior that could cause problems.

Resolves #1914.