warehouses/postgresql: remove test file
We currently have the file 'updates_sql_test.go' that tests the
correctness (just a little) of the file containing the warehouse
updates.

This test has been skipped because it is problematic to test the current
contents of the PostgreSQL updates.sql file (it contains a 'RENAME'
statement that fails when executed on a database already updated).

Furthermore, considering the limited utility of this file, and that it
is a temporary solution for a real update system, this commit is
removing this test file because it adds more complexity than it
provides.