core/internal/transformers: clarify errors in local transformations
Previously, when running a local transformation function, any failure in
the interpreter process produced an unclear log message. The message
often made it difficult to identify the root cause or even to understand
that the issue was related to a transformation function.

This change makes such errors explicit and informative.

Fixes #1940