assets: remove strategy selection when creating a new connection
When adding a source connection such as JavaScript or Android, the
strategy selection menu was displayed. However, the concept of
"strategy" is too advanced for users adding a connection for the first
time. They might either stop the process or simply keep the default
option.

This change removes the strategy selection step during connection
creation, leaving it available only in the settings.