core/internal/transformers/lambda: eval argument in clean environment
Currently, the evaluation of the Lambda argument 'event', which contains
the actual data to be transformed, occurs in the same environment in
which 'eval' is called.

This command ensures that the evaluation occurs in a controlled and
minimal environment, to avoid problems now and in the future (e.g.,
conflicts with overridden or removed names in the wrapper).