core/internal/schemas: remove unnecessary comment from event schema
While the comment itself is correct, such a comment should be added to
many other fields as well.

Beyond this, unlike the initial development phase where certain aspects
were unclear, the functioning of "muid" is now more defined and the
comment is not as informative.

For these reasons, this commit removes this comment.