assets: add manual reload of samples
This commit adds the ability to manually reload samples in the actions full
mode.

This new feature is now available for all actions that support samples, except
for those that handle events and therefore use the event listener to retrieve
them.

In addition, this commit makes the first batch of samples for database import
actions load automatically when full mode is opened, removing the need for
manual confirmation via button press.

Closes #1881