core/internal/transformers: fix transf. function for Python < 3.13
In Python's 'exec' builtin, the 'globals' parameter can be passed as a
key-argument only since Python 3.13. From the Python documentation:

	Changed in version 3.13: The globals and locals arguments can now be
	passed as keywords.

This means that on versions of Python prior to 3.13, execution of the
transformation function wrapper fails, returning an error.

This commit therefore removes the parameter name, passing 'globals' as
positional. The behavior is completely identical, but this makes the
wrapper work even on older versions of Python.