commit: run the tests in Go after doing the other operations
Currently, in the commit script, operations such as admin changes are
performed after running the Go tests.

This commit ensures that all operations that could modify the repository
are performed first, and only at the end does it run the Go tests, which
are then run on a "final" repository, already cleaned up and modified as
needed.