cmd: stop writing to `error.log` and `panics.log`
This commit ensures that all Meergo errors and panics are written
exclusively to stderr, rather than being written to files.

The system that starts Meergo will then decide what and where to save
errors.

This is useful because it makes using Meergo simpler and more flexible,
as this change no longer requires the writing permissions to create the
two aforementioned files.