all: use NODEJS instead of NODE for transformer env vars
The variable name MEERGO_TRANSFORMERS_LOCAL_NODE_EXECUTABLE could be
misread as "the executable of the local node." and furthermore while
NODE might be acceptable when referring strictly to the binary name, it
becomes incorrect in MEERGO_TRANSFORMERS_AWS_LAMBDA_NODE_RUNTIME, where
the intent is to reference the Node.js framework, not the executable.

This commit replaces NODE with NODEJS in both the local transformer and
the AWS Lambda transformer to ensure clarity and avoid ambiguity.