cmd, admin: add env variable for URL of potential connectors JSON file
This commit introduces the environment variable:

    MEERGO_POTENTIAL_CONNECTORS_URL

which allows to specify where to get the JSON file for connectors.

This is particularly useful in cases where a change is made to Meergo
via a PR that modifies the JSON file, a PR review is requested, and the
reviewer user wants to be able to test reading a different file without
having to modify a versioned file, as is currently the case with 'main',
where it is necessary to modify the constant in the Admin.

This commit also adds the option to not import the JSON file entirely,
useful in cases where you don't want to retrieve a file from the network
when running Meergo.

Resolves #2003.

Co-authored-by: Andrea Retini <retini@open2b.com>