meergo: restart Meergo if it crashes when running with Docker Compose
Currently, if the Meergo executable crashes while running the Docker
Compose "meergo" service automatically, it is not restarted and the user
must manually intervene in Docker Compose.

This change ensures that if Meergo fails (e.g., a panic due to a bug in
a specific part of the software), Docker will restart the service again,
allowing the user to perhaps continue browsing other Admin pages that
are not affected by the bug or perform other operations.