all: move PostgreSQL database initialization within Meergo
Currently, you need to manually run the scripts to initialize the
PostgreSQL database.

This also makes it difficult to use via Docker Compose, because the
Meergo image does not contain these scripts, and therefore, for a given
Meergo image, the corresponding initialization scripts must be used,
which is not always simple or straightforward.

For these reasons, this commit moves database initialization into
Meergo; this can be activated via command-line options. Initialization
of Docker-specific members has also been moved into the executable.

Resolves #2005.