admin,core: allow `null`/`None` values in transformed properties
Previously, when running a JavaScript or Python transformation during
user import, passing `null` or `None` as a property value was not
allowed. This made transformation functions unnecessarily complex: they
had to check not only whether a property existed, but also whether it
wasn't `null` or `None`.

With this commit, if a transformed property ends up being `null` or
`None`, it is simply treated as absent and no error is returned.