all: support `doas` in addition to `sudo`
The Docker image of Meergo is built upon Alpine Linux. Their
documentation recommend the use of 'doas' instead of 'sudo', because it
is simpler and lighter.

Using 'doas' fixes the issue #2019.

The support for 'sudo' must be kept, because it is the most frequently
used option in systems that run Meergo; so, we have two different
environment variables, which can be set alternatively, one for using
'sudo', one for using 'doas'.

Resolves #2019.