core/internal/datastore/diffschemas: improve validations on rePaths
Currently, the function that diffs two schemas accepts cases where the
rePaths indicate inconsistent information with respect to the properties
of the old and new schemas.

This can lead to inconsistencies, which in turn generate inconsistent
queries. See issue #2026.

For this reason, this commit handles such inconsistencies by returning
errors that indicate the problem. It also adds tests on these cases.

Resolves #2026.