core: omit `linkedConnection` when events aren't supported
Make the behavior of 'linkedConnection' consistent with the other keys
whose presence depends on a connection's targets and role.

Previously, the API returned 'linkedConnection' as 'null' when the
connection didn't support events and as a non-null value when it did.
Now, if the connection doesn't support events, 'linkedConnection' is
simply not returned.