core/internal/datastore: make name of env var used in tests explicit
Currently, the name of the environment variable containing the Snowflake
configuration is not explicitly coded in the test because it is
constructed dynamically.

Considering that there is no reason to construct it dynamically, given
that the code in the test is already specific to Snowflake, and that the
variable doesn't appear when searching in Meergo, making it difficult to
identify the tests using it, this commit makes the variable name
explicit.