all: pin specific third-party software versions and related changes
This commit changes the versions of various third-party software that
Meergo depends on (in use and testing); in general, the approach adopted
is to pin the version up to and including the minor version, excluding
patches. This appears to be the best compromise, at the moment, between
these two requirements:

1. Software updated with the latest security fixes
2. Software that behaves reproducibly, thus not silently introducing
   incompatibilities and/or new features.

For this reason, the versions of various third-party software have been
updated and specified.

Furthermore, dependencies have been reduced (for example, by explicitly
setting the Alpine Linux operating system, without delegating the choice
of base OS to Docker images) to reduce the need for security updates and
issues.